+++ test/resources/example.tsv
@@ -1,0 +1,13 @@
+#FORMAT=WebAnno TSV 3.3
+#T_SP=Layer1|Field1
+#T_SP=Layer2|Field2|Field3
+#T_RL=Layer3|Field4|BT_Layer2
+
+
+#Text=First sentence
+1-1	0-5	First	_	*	UVW	_	_
+1-2	6-14	sentence	ABC	*[1]	XYZ[1]	R	1-1[0_1]
+
+#Text=Second sentence
+2-1	15-21	Second	_	*[1]	XYZ[1]	_	_
+2-2	22-30	sentence	DEF	_	_	_	_
