+++ webanno_tsv/__init__.py
@@ -2,22 +2,10 @@
 from .webanno_tsv import (
     Token,
     Sentence,
-    AnnotationPart,
     SpanAnnotation,
     RelationAnnotation,
-    SpanLayer,
-    RelationLayer,
+    SpanLayerDefinition,
+    RelationLayerDefinition,
+    Layer,
     Document,
-
-    NO_LABEL_ID,
-
-    fix_annotation_ids,
-    merge_into_annotations,
-    token_sort,
-    tokens_from_strs,
-    utf_16_length,
-
-    webanno_tsv_write,
-    webanno_tsv_read_file,
-    webanno_tsv_read_string,
 )
